+++ Source/CesiumRuntime/Public/Cesium3DTileset.h
@@ -27,6 +27,10 @@
 #include <vector>
 #include "Cesium3DTileset.generated.h"
 
+#ifdef CESIUM_DEBUG_TILE_STATES
+#include <Cesium3DTilesSelection/DebugTileStateDatabase.h>
+#endif
+
 class UMaterialInterface;
 class ACesiumCartographicSelection;
 class ACesiumCameraManager;
@@ -1096,8 +1100,10 @@ public:
    * This method is not supposed to be called by clients. It is currently
    * only required by the UnrealResourcePreparer.
    *
+   * @internal
    * See {@link
-   * Cesium3DTilesetRoot::GetCesiumTilesetToUnrealRelativeWorldTransform}.
+   * UCesium3DTilesetRoot::GetCesiumTilesetToUnrealRelativeWorldTransform}.
+   * @endinternal
    */
   const glm::dmat4& GetCesiumTilesetToUnrealRelativeWorldTransform() const;
 
@@ -1237,6 +1243,10 @@ private:
 
 private:
   TUniquePtr<Cesium3DTilesSelection::Tileset> _pTileset;
+
+#ifdef CESIUM_DEBUG_TILE_STATES
+  TUniquePtr<Cesium3DTilesSelection::DebugTileStateDatabase> _pStateDebug;
+#endif
 
   std::optional<FCesiumFeaturesMetadataDescription>
       _featuresMetadataDescription;