+++ Source/CesiumRuntime/Public/CesiumCommon.h
@@ -8,5 +8,3 @@
   (ENGINE_MAJOR_VERSION > 5 || ENGINE_MINOR_VERSION >= 5)
 #define ENGINE_VERSION_5_4_OR_HIGHER                                           \
   (ENGINE_MAJOR_VERSION > 5 || ENGINE_MINOR_VERSION >= 4)
-#define ENGINE_VERSION_5_3_OR_HIGHER                                           \
-  (ENGINE_MAJOR_VERSION > 5 || ENGINE_MINOR_VERSION >= 3)