+++ Source/CesiumRuntime/Public/CesiumFeatureIdAttribute.h
@@ -28,8 +28,9 @@ enum class ECesiumFeatureIdAttributeStatus : uint8 {
 
 /**
  * @brief A blueprint-accessible wrapper for a feature ID attribute from a glTF
- * primitive. Provides access to per-vertex feature IDs which can be used with
- * the corresponding {@link FCesiumFeatureTable} to access per-vertex metadata.
+ * model. Provides access to feature IDs which can be used with the
+ * corresponding {@link FCesiumPropertyTable} to access metadata. These feature
+ * IDs may be defined per-vertex or per-instance.
  */
 USTRUCT(BlueprintType)
 struct CESIUMRUNTIME_API FCesiumFeatureIdAttribute {
@@ -109,7 +110,6 @@ public:
   UFUNCTION(
       BlueprintCallable,
       BlueprintPure,
-      Category = "Cesium|Metadata|FeatureIdAttribute",
       Meta =
           (DeprecatedFunction,
            DeprecationMessage =
@@ -131,27 +131,35 @@ public:
       UPARAM(ref) const FCesiumFeatureIdAttribute& FeatureIDAttribute);
 
   /**
-   * Get the number of vertices in the primitive containing the feature
-   * ID attribute. If the feature ID attribute is invalid, this returns 0.
+   * Gets the number of elements in the attribute. This is distinct from the
+   * number of unique feature IDs within the attribute.
+   *
+   * For a feature ID attribute of a regular mesh, this is the number of
+   * vertices. For a per-instance feature ID, this is the number of instances.
+   *
+   * If the feature ID attribute is invalid, this returns 0.
    */
   UFUNCTION(
       BlueprintCallable,
       BlueprintPure,
       Category = "Cesium|Features|FeatureIDAttribute")
   static int64
-  GetVertexCount(UPARAM(ref)
-                     const FCesiumFeatureIdAttribute& FeatureIDAttribute);
+  GetCount(UPARAM(ref) const FCesiumFeatureIdAttribute& FeatureIDAttribute);
 
   /**
-   * Gets the feature ID associated with the given vertex. The feature ID can be
-   * used with a FCesiumFeatureTable to retrieve the per-vertex metadata. If
-   * the feature ID attribute is invalid, this returns -1.
+   * Gets the feature ID at the given index. A feature ID can be used with a
+   * FCesiumPropertyTable to retrieve the metadata for that ID. If the feature
+   * ID attribute is invalid, this returns -1.
+   *
+   * For a feature ID attribute of a regular mesh, the provided Index is the
+   * index of a vertex within the mesh. For a per-instance feature ID, the
+   * provided Index is the index of the instance.
    */
   UFUNCTION(
       BlueprintCallable,
       BlueprintPure,
       Category = "Cesium|Features|FeatureIDAttribute")
-  static int64 GetFeatureIDForVertex(
+  static int64 GetFeatureID(
       UPARAM(ref) const FCesiumFeatureIdAttribute& FeatureIDAttribute,
-      int64 VertexIndex);
+      int64 Index);
 };